# src/v1/items/dto/create-item.dto.ts
import {
  IsDateString,
  IsNotEmpty,
  IsNumber,
  IsPositive,
} from 'class-validator';

export class CreateItemDto {
  @IsNotEmpty()
  name: string;

  @IsNotEmpty()
  @IsNumber()
  @IsPositive()
  startPrice: number;

  @IsNotEmpty()
  @IsDateString()
  endedAt: Date;
}